refactor: 백엔드와 말하는 코드를 분리한다 (#117)
main.py 1,224 -> 1,023줄. backend_client.py 244줄.

왜 이것만 뗐나
  계약 시험 23개로 덮여 있고 경계가 명확하며
  나머지 코드와 공유하는 상태가 없다.

왜 나머지는 안 뗐나
  summarize_text_auto 가 471줄인데 테스트가 0개다.
  테스트 없이 쪼개면 '동작이 바뀌었는지 알 방법이 없는 변경' 이 되고,
  471줄이라 테스트를 쓸 수도 없다 - 닭-달걀이다.

  이 저장소의 규칙은 '모든 변경에 전/후를 붙인다' 이고
  여기서는 지킬 수 없다. 먼저 시험 가능한 크기로 만드는 작업이
  필요하고 그건 별개다.

  '정리했습니다' 로 전부 옮기는 것보다 뗄 수 있는 것만 떼고
  나머지를 왜 안 뗐는지 적는 편이 정직하다.

분리 후 확인
  시험이 통과하는 것만으로는 부족하다. 시험은 main 을 import 하므로
  옛 코드를 그대로 보고 있을 수도 있다.

    main.send_summary_to_api is backend_client.send_summary_to_api  # True

  같은 객체인지 확인하고 backend_client.py 를 고쳐 2개가 깨지는 것까지 봤다.
