+++ ai/README.md
@@ -131,3 +131,49 @@ GET /
 - [ ] 더 정교한 한국어 형태소 분석기 연동
 - [ ] 성능 모니터링 대시보드
 - [ ] 다양한 임베딩 모델 지원
+
+---
+
+## 구조 (2026-08-25, #117)
+
+```
+main.py              엔드포인트 · 오디오 병합 · STT · 요약
+backend_client.py    백엔드(Java)와 말하는 코드
+tests/               계약 시험 23개
+```
+
+### 왜 이것만 분리했나
+
+`backend_client.py` 로 옮긴 함수들은
+
+- **계약 시험 23개로 덮여 있다** (`contracts/internal-api.json` 기준)
+- 경계가 명확하다 — 백엔드와 말하는 일만 한다
+- 나머지 코드와 공유하는 상태가 없다
+
+### 왜 나머지는 안 뗐나
+
+```
+summarize_text_auto   471줄   테스트 0개
+merge_audio           168줄
+Start_STT             142줄
+```
+
+**테스트 없이 471줄 함수를 쪼개면 "동작이 바뀌었는지 알 방법이 없는 변경" 이 된다.**
+그런데 471줄이라 테스트를 쓸 수도 없다 — 닭-달걀이다.
+
+이 저장소의 규칙은 *"모든 변경에 전/후를 붙인다"* 이고 여기서는 지킬 수 없다.
+먼저 그 함수를 **시험 가능한 크기로 만드는 작업**이 필요하고, 그건 별개다.
+
+> "정리했습니다" 로 전부 옮기는 것보다, 뗄 수 있는 것만 떼고
+> **나머지를 왜 안 뗐는지 적는 편이 정직하다.**
+
+### 분리 후 확인한 것
+
+시험이 그대로 통과하는 것만으로는 부족하다 — 시험은 `main` 을 import 하므로
+**옛 코드를 그대로 보고 있을 수도 있다.**
+
+```python
+main.send_summary_to_api is backend_client.send_summary_to_api   # True
+```
+
+같은 객체인지 확인하고, `backend_client.py` 를 고쳐서 시험이 깨지는 것까지 봤다.
